+++ src/Enums/Type.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Khamsolt\Orchid\Files\Enums;
+
+enum Type: string
+{
+    case IMAGES = 'images';
+
+    case FILES = 'files';
+
+    case AUDIO = 'audio';
+
+    case VIDEO = 'video';
+}