+++ tests/TestCase.php
@@ -3,8 +3,6 @@
 namespace Khamsolt\Orchid\Files\Tests;
 
 use Artisan;
-use Illuminate\Database\Eloquent\Factories\Factory;
-use Illuminate\Database\Schema\Blueprint;
 use Khamsolt\Orchid\Files\FileGenerator;
 use Khamsolt\Orchid\Files\FileServiceProvider;
 use Orchestra\Testbench\TestCase as Orchestra;
@@ -12,46 +10,21 @@ use Orchid\Platform\Providers\FoundationServiceProvider;
 
 class TestCase extends Orchestra
 {
+    protected $loadEnvironmentVariables = false;
+
     protected function setUp(): void
     {
         parent::setUp();
-
-        Factory::guessFactoryNamesUsing(
-            fn (string $modelName) => 'Khamsolt\\Orchid\\Files\\Database\\Factories\\'.class_basename($modelName).'Factory'
-        );
-
-        $this->setUpDatabase($this->app);
-        $this->startSession();
-    }
-
-    /**
-     * Set up the database.
-     *
-     * @param \Illuminate\Foundation\Application $app
-     */
-    protected function setUpDatabase($app)
-    {
-        $app['db']->connection()->getSchemaBuilder()->create('users', function (Blueprint $table) {
-            $table->id();
-            $table->string('name');
-            $table->string('email')->unique();
-            $table->timestamp('email_verified_at')->nullable();
-            $table->string('password');
-            $table->rememberToken();
-            $table->timestamps();
-        });
-
-        $this->loadMigrationsFrom(dirname(__DIR__) . '/vendor/orchid/platform/database/migrations');
 
         Artisan::call('vendor:publish', [
             '--provider' => FoundationServiceProvider::class,
             '--tag' => 'config',
         ]);
 
-        Artisan::call('migrate', ['--force' => true]);
+        Artisan::call('orchid-files:install');
     }
 
-    protected function getPackageProviders($app)
+    protected function getPackageProviders($app): array
     {
         return [
             FoundationServiceProvider::class,
@@ -59,16 +32,15 @@ class TestCase extends Orchestra
         ];
     }
 
-    /**
-     * Set up the environment.
-     *
-     * @param \Illuminate\Foundation\Application $app
-     */
-    protected function getEnvironmentSetUp($app)
+    protected function defineDatabaseMigrations(): void
     {
-        //$app['path.lang'] = __DIR__ . '/lang';
+        $this->loadLaravelMigrations();
 
-        $app['config']->set('database.default', 'pgsql');
+        $this->beforeApplicationDestroyed(fn() => $this->artisan('db:wipe'));
+    }
+
+    protected function getEnvironmentSetUp($app): void
+    {
         $app['config']->set('database.connections.pgsql', [
             'driver' => 'pgsql',
             'host' => 'postgres-postgis',
@@ -78,15 +50,5 @@ class TestCase extends Orchestra
         ]);
 
         $app['config']->set('platform.attachment.generator', FileGenerator::class);
-    }
-
-    /**
-     * @inheritdoc
-     */
-    protected function resolveApplicationConfiguration($app)
-    {
-        parent::resolveApplicationConfiguration($app);
-
-        $app['config']->set('session.drive', 'array');
     }
 }